Disallow infinite depagination recursion.
Should infinite recursion be a sign that some pages are faulty?
Which ones?

For the record, the website that triggered the need for restrictions
on depagination was http://www.lifehacker.com since that site now
has a link whose text reads "Learn More &raquo;"
[they had a redesign to which that text refers]; "&raquo;" is the
"Right Angle QUOte" that is present in one of the regexes.